Create chat and text completion extension traits
